# backend/openapi.yaml
openapi: 3.0.0
info:
  title: AI Flirt Studio API
  description: Backend API for AI Flirt Translator - Credits & Subscription Engine
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:3001/api
    description: Development server
  - url: https://api.example.com/api
    description: Production server
...
paths:
...
  /admin/users:
    get:
      summary: List users
      tags: [Admin]
      security:
        - apiKeyAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of users
  /admin/adjust_credits:
    post:
      summary: Adjust user credits
      tags: [Admin]
      security:
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user_id
                - amount
              properties:
                user_id:
                  type: string
                  format: uuid
                amount:
                  type: integer
                reason:
                  type: string
      responses:
        '200':
          description: Credits adjusted
...